爆弾が0個のときに周りも一緒に開くようにする
